Drop setuptools_scm
It seems that setuptools_scm doesn't work well with pre-commit based
installations where there isn't either a built package or a git
repository to get the version info from.

As well, the version needs to be incremented in the README, so it makes
sense to switch to manually setting the version info.